Revert "change phishlabsioc to phishlabsioc_eir"
This reverts commit 01c5c369f3c9dc38d3d6e050a467beb66aada563.
