+++ Integrations/PhishLabsIOC/CHANGELOG.md
@@ -1,0 +1,5 @@
+## [Unreleased]
+
+
+## [19.8.0] - 2019-08-06
+  - Fixed an issue with the **updatedAt** field.
