+++ Integrations/PhishLabsIOC_EIR/CHANGELOG.md
@@ -1,6 +1,6 @@
 ## [Unreleased]
-- change name to PhishLabsIOC_EIR from PhishLabsIOC
 
-## [19.8.0] - 2019-08-06
-  - Fixed an issue with the **updatedAt** field.
 
+## [19.11.1] - 2019-11-26
+#### New Integration
+Get live feeds of IOC data from PhishLabs.
